client/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
numpy==1.23.2
pandas==1.4.3
progressbar2==4.0.0
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.11